+++ changelog.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+- Update gregoryv/english dependency
+
 ## [0.22.0] 2021-10-23
 
 - Add theme GoishColors